requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
scipy
azure-storage-blob
google-cloud-storage